requirements.txt
alembic==0.8.10
aniso8601==3.0.0
arrow==0.12.1
blinker==1.4
click==6.7
coverage==4.5.1
factory-boy==2.9.2
Faker==0.8.13
Flask==0.12
Flask-Migrate==2.0.3
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.0
psycopg2==2.7.3.1
py==1.5.3
pytest==3.0.6
pytest-cov==2.5.1
python-dateutil==2.7.2
python-dotenv==0.6.3
python-editor==1.0.3
pytz==2018.4
raven==6.6.0
redis==2.10.5
six==1.11.0
SQLAlchemy==1.1.5
text-unidecode==1.2
Werkzeug==0.14.1
